% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_controls.R
\name{verify_inputs}
\alias{verify_inputs}
\title{Verify the inputs to \code{\link{optimize_controls}}}
\usage{
verify_inputs(X, importances, ratio, q_s, st, z, integer, solver)
}
\arguments{
\item{X}{a matrix or data frame containing constraints in the columns. The number
of rows should equal the length of \code{z}. Balance is achieved when a constraint
sums to 0, such that numbers closer to 0 are better. When a constraint
does not apply to a particular unit, the entry should be \code{NA}.
This should typically be generated using \code{\link{generate_constraints}}.}

\item{importances}{a vector with length equal to the number of constraints or columns
in \code{X}. This can be generated using \code{\link{generate_constraints}} and each nonnegative value
denotes how much to prioritize each constraint, with the default being 1
for all constraints.}

\item{ratio}{a numeric specifying the desired ratio of controls to treated in
each stratum. If \code{NULL}, \code{q_s} should be specified.}

\item{q_s}{a named vector indicating how many control units are to be selected from each stratum.
If \code{NULL}, \code{ratio} should be specified. If both are specified, \code{q_s} will take priority.
Typically, if the desired ratio is not feasible for every stratum, \code{q_s} should be generated
using \code{\link{generate_qs}}.}

\item{st}{a stratum vector with the \code{i}th entry equal to the
stratum of unit \code{i}. This should have the same order of units and length
as \code{z}.}

\item{z}{a treatment indicator vector with \code{i}th entry equal to 0 if
unit \code{i} is a control and equal to 1 if unit \code{i} is treated.}

\item{integer}{a logical stating whether to use a mixed integer programming solver
instead of randomized rounding. Default is \code{FALSE}.}

\item{solver}{a character stating which solver to use to run the linear program.
Options are "Rglpk" (default) or "gurobi". You must have gurobi installed to use "gurobi".
If gurobi is available, this is the recommended solver.}
}
\description{
Makes sure that the inputs to \code{\link{optimize_controls}} are in the correct
format and feasible.
}
\keyword{internal}
